requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
defusedxml==0.8.0rc2
Django==5.0.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
ecdsa==0.18.0
exceptiongroup==1.2.0
Faker==22.6.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
